:recycle: Refactor: 회원 탈퇴 API 위치 변경 (#14) - member -> myPage
