+++ src/main/java/forus/naviforyou/domain/member/controller/MemberController.java
@@ -9,12 +9,10 @@ import forus.naviforyou.domain.member.service.KakaoService;
 import forus.naviforyou.domain.member.service.MemberService;
 import forus.naviforyou.domain.member.service.NaverService;
 import forus.naviforyou.global.common.BaseResponse;
-import forus.naviforyou.global.common.collection.member.Member;
 import forus.naviforyou.global.error.dto.ErrorCode;
 import forus.naviforyou.global.error.exception.BaseException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -102,15 +100,6 @@ public class MemberController {
         return BaseResponse.ok(tokenRes);
     }
 
-    @PostMapping("/delete")
-    private ResponseEntity<?> delete(@RequestBody DeleteReq deleteReq, @AuthenticationPrincipal Member member){
-        memberService.delete(deleteReq.getPwd(), member);
-        return BaseResponse.ok(
-                DeleteRes.builder()
-                        .result(true)
-                        .build()
-        );
-    }
 
     @GetMapping("/searchPassword/sendCode")
     public ResponseEntity<?> searchPwdChkCode(@RequestParam String email){
