+++ src/main/java/forus/naviforyou/domain/myPage/dto/reponse/ResultRes.java
@@ -1,0 +1,7 @@
+package forus.naviforyou.domain.myPage.dto.reponse;
+
+import lombok.Getter;
+
+@Getter
+public record ResultRes(Boolean result) {
+}
